requirements.txt
REQUIREMENTS :-
1.opencv-python
2.pillow
3.pandas
4.numpy

